- Added bit writer as wrapper around byte writer  - Added small ints, unary encoding and rice encoding to bit writer

+++ HandySerialization/Extensions/Vectors.cs
@@ -215,7 +215,7 @@ public static class Vectors
     {
         return new Complex(
             reader.ReadFloat64(),
-            reader.ReadFloat32()
+            reader.ReadFloat64()
         );
     }
